fix: 등록한 아이디를 기준으로 DB에 있는 작품을 내 서재에서 삭제 가능한 것 수정

+++ controller/series_controller.go
@@ -545,6 +545,14 @@ func (controller SeriesController) DeleteUserRecordEmptySeries(ctx *fiber.Ctx) e
 		})
 	}
 
+	if recordEntity.SeriesId != 0 {
+		return ctx.Status(fiber.StatusBadRequest).JSON(model.GeneralResponse{
+			Code:    fiber.StatusBadRequest,
+			Message: "This is not user added empty series",
+			Data:    nil,
+		})
+	}
+
 	err = controller.UserRecordSeriesService.DeleteUserRecordSeriesByUserIdAndId(ctx.Context(), userEntity.Id, request.Id)
 	if err != nil {
 		return ctx.Status(fiber.StatusInternalServerError).JSON(model.GeneralResponse{
